Adjust the useApi hook to cancel previous requests
