+++ src/util/call-api.js
@@ -1,17 +1,9 @@
-import _ from 'lodash';
-import { flow, join, keys, map, omitBy } from 'lodash/fp';
 import axios from 'axios';
 
+import buildApiUrl from './build-api-url';
+
 const callApi = async (method, params, opts) => {
-  const options = _.defaults({}, opts, { version: 1 });
-  const endpoint = process.env.REACT_APP_API_ENDPOINT;
-  const querystring = flow([
-    omitBy(value => value === undefined),
-    keys,
-    map(key => `${key}=${params[key]}`),
-    join('&'),
-  ])(params);
-  const url = `${endpoint}/v${options.version}/${method}?${querystring}`;
+  const url = buildApiUrl(method, params, opts);
   const response = await axios.get(url);
 
   return response.data;
